Address comment

+++ src/main/java/org/stratumproject/fabric/tna/behaviour/upf/FabricUpfProgrammable.java
@@ -83,32 +83,43 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
     private ApplicationId appId;
 
     @Override
+    protected boolean setupBehaviour(String opName) {
+        // Already initialized.
+        if (appId != null) {
+            return true;
+        }
+
+        if (!super.setupBehaviour(opName)) {
+            return false;
+        }
+        flowRuleService = handler().get(FlowRuleService.class);
+        packetService = handler().get(PacketService.class);
+        fabricUpfStore = handler().get(DistributedFabricUpfStore.class);
+        upfTranslator = new FabricUpfTranslator(fabricUpfStore);
+        final CoreService coreService = handler().get(CoreService.class);
+        appId = coreService.getAppId(PipeconfLoader.APP_NAME);
+        if (appId == null) {
+            log.warn("Application ID is null. Cannot initialize behaviour.");
+            return false;
+        }
+
+        var capabilities = new FabricCapabilities(pipeconf);
+        if (!capabilities.supportUpf()) {
+            log.warn("Pipeconf {} on {} does not support UPF capabilities, " +
+                             "cannot perform {}",
+                     pipeconf.id(), deviceId, opName);
+            return false;
+        }
+        return true;
+    }
+
+    @Override
     public boolean init() {
-        if (super.setupBehaviour("init()")) {
+        if (setupBehaviour("init()")) {
             if (!computeHardwareResourceSizes()) {
                 // error message will be printed by computeHardwareResourceSizes()
                 return false;
             }
-
-            flowRuleService = handler().get(FlowRuleService.class);
-            packetService = handler().get(PacketService.class);
-            fabricUpfStore = handler().get(DistributedFabricUpfStore.class);
-            upfTranslator = new FabricUpfTranslator(fabricUpfStore);
-            final CoreService coreService = handler().get(CoreService.class);
-            appId = coreService.getAppId(PipeconfLoader.APP_NAME);
-            if (appId == null) {
-                log.warn("Application ID is null. Cannot initialize behaviour.");
-                return false;
-            }
-
-            var capabilities = new FabricCapabilities(pipeconf);
-            if (!capabilities.supportUpf()) {
-                log.warn("Pipeconf {} on {} does not support UPF capabilities, " +
-                                "cannot perform {}",
-                        pipeconf.id(), deviceId, "init()");
-                return false;
-            }
-
             log.info("UpfProgrammable initialized for appId {} and deviceId {}", appId, deviceId);
             return true;
         }
@@ -170,6 +181,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void enablePscEncap(int defaultQfi) {
+        if (!setupBehaviour("enablePscEncap()")) {
+            return;
+        }
         if (pipeconf.pipelineModel().table(FABRIC_EGRESS_SPGW_GTPU_ENCAP).isEmpty()) {
             log.error("Missing {} table in {}, cannot enable PSC encap",
                       FABRIC_EGRESS_SPGW_GTPU_ENCAP, deviceId);
@@ -181,6 +195,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void disablePscEncap() {
+        if (!setupBehaviour("disablePscEncap()")) {
+            return;
+        }
         if (pipeconf.pipelineModel().table(FABRIC_EGRESS_SPGW_GTPU_ENCAP).isEmpty()) {
             log.debug("Missing {} table in {}, assuming PSC encap is disabled by default",
                       FABRIC_EGRESS_SPGW_GTPU_ENCAP, deviceId);
@@ -192,6 +209,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void sendPacketOut(ByteBuffer data) {
+        if (!setupBehaviour("sendPacketOut()")) {
+            return;
+        }
         final OutboundPacket pkt = new DefaultOutboundPacket(
                 deviceId,
                 // Use TABLE logical port to have pkt routed via pipeline tables.
@@ -204,6 +224,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void cleanUp() {
+        if (!setupBehaviour("cleanUp()")) {
+            return;
+        }
         log.info("Clearing all UPF-related table entries.");
         flowRuleService.removeFlowRulesById(appId);
         fabricUpfStore.reset();
@@ -211,6 +234,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void clearInterfaces() {
+        if (!setupBehaviour("clearInterfaces()")) {
+            return;
+        }
         log.info("Clearing all UPF interfaces.");
         for (FlowRule entry : flowRuleService.getFlowEntriesById(appId)) {
             if (upfTranslator.isFabricInterface(entry)) {
@@ -230,6 +256,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void clearFlows() {
+        if (!setupBehaviour("clearFlows()")) {
+            return;
+        }
         log.info("Clearing all UE sessions.");
         int pdrsCleared = 0;
         int farsCleared = 0;
@@ -248,6 +277,10 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public Collection<PdrStats> readAllCounters(long maxCounterId) {
+        if (!setupBehaviour("readAllCounters()")) {
+            return null;
+        }
+
         long counterSize = pdrCounterSize();
         if (maxCounterId != -1) {
             counterSize = Math.min(maxCounterId, counterSize);
@@ -306,21 +339,33 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public long pdrCounterSize() {
+        if (!setupBehaviour("pdrCounterSize()")) {
+            return -1;
+        }
         return pdrCounterSize;
     }
 
     @Override
     public long farTableSize() {
+        if (!setupBehaviour("farTableSize()")) {
+            return -1;
+        }
         return farTableSize;
     }
 
     @Override
     public long pdrTableSize() {
+        if (!setupBehaviour("pdrTableSize()")) {
+            return -1;
+        }
         return Math.min(encappedPdrTableSize, unencappedPdrTableSize) * 2;
     }
 
     @Override
     public PdrStats readCounter(int cellId) throws UpfProgrammableException {
+        if (!setupBehaviour("readCounter()")) {
+            return null;
+        }
         if (cellId >= pdrCounterSize() || cellId < 0) {
             throw new UpfProgrammableException("Requested PDR counter cell index is out of bounds.",
                     UpfProgrammableException.Type.COUNTER_INDEX_OUT_OF_RANGE);
@@ -364,6 +409,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void addPdr(PacketDetectionRule pdr) throws UpfProgrammableException {
+        if (!setupBehaviour("addPdr()")) {
+            return;
+        }
         if (pdr.counterId() >= pdrCounterSize() || pdr.counterId() < 0) {
             throw new UpfProgrammableException("Counter cell index referenced by PDR is out of bounds.",
                     UpfProgrammableException.Type.COUNTER_INDEX_OUT_OF_RANGE);
@@ -377,6 +425,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void addFar(ForwardingActionRule far) throws UpfProgrammableException {
+        if (!setupBehaviour("addFar()")) {
+            return;
+        }
         FlowRule fabricFar = upfTranslator.farToFabricEntry(far, deviceId, appId, DEFAULT_PRIORITY);
         log.info("Installing {}", far.toString());
         flowRuleService.applyFlowRules(fabricFar);
@@ -385,6 +436,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void addInterface(UpfInterface upfInterface) throws UpfProgrammableException {
+        if (!setupBehaviour("addInterface()")) {
+            return;
+        }
         FlowRule flowRule = upfTranslator.interfaceToFabricEntry(upfInterface, deviceId, appId, DEFAULT_PRIORITY);
         log.info("Installing {}", upfInterface);
         flowRuleService.applyFlowRules(flowRule);
@@ -426,6 +480,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public Collection<PacketDetectionRule> getPdrs() throws UpfProgrammableException {
+        if (!setupBehaviour("getPdrs()")) {
+            return null;
+        }
         ArrayList<PacketDetectionRule> pdrs = new ArrayList<>();
         for (FlowRule flowRule : flowRuleService.getFlowEntriesById(appId)) {
             if (upfTranslator.isFabricPdr(flowRule)) {
@@ -437,6 +494,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public Collection<ForwardingActionRule> getFars() throws UpfProgrammableException {
+        if (!setupBehaviour("getFars()")) {
+            return null;
+        }
         ArrayList<ForwardingActionRule> fars = new ArrayList<>();
         for (FlowRule flowRule : flowRuleService.getFlowEntriesById(appId)) {
             if (upfTranslator.isFabricFar(flowRule)) {
@@ -448,6 +508,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public Collection<UpfInterface> getInterfaces() throws UpfProgrammableException {
+        if (!setupBehaviour("getInterfaces()")) {
+            return null;
+        }
         ArrayList<UpfInterface> ifaces = new ArrayList<>();
         for (FlowRule flowRule : flowRuleService.getFlowEntriesById(appId)) {
             if (upfTranslator.isFabricInterface(flowRule)) {
@@ -459,6 +522,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void removePdr(PacketDetectionRule pdr) throws UpfProgrammableException {
+        if (!setupBehaviour("removePdr()")) {
+            return;
+        }
         final PiCriterion match;
         final PiTableId tableId;
         if (pdr.matchesEncapped()) {
@@ -479,6 +545,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void removeFar(ForwardingActionRule far) throws UpfProgrammableException {
+        if (!setupBehaviour("removeFar()")) {
+            return;
+        }
         log.info("Removing {}", far.toString());
 
         PiCriterion match = PiCriterion.builder()
@@ -490,6 +559,9 @@ public class FabricUpfProgrammable extends AbstractP4RuntimeHandlerBehaviour
 
     @Override
     public void removeInterface(UpfInterface upfInterface) throws UpfProgrammableException {
+        if (!setupBehaviour("removeInterface()")) {
+            return;
+        }
         Ip4Prefix ifacePrefix = upfInterface.getPrefix();
         if (upfInterface.isCore()) {
             applyUplinkRecirculation(ifacePrefix, true);
